consolidate imports and enhance TextfieldToolSettings with multiline support
- Merged import statements for ToolSetting and JsonSchema from '@formswizard/types'.
- Added multiline property to the TextfieldToolSettings schema.
- Updated mapping functions to handle multiline option in tool data and wizard UI schema.
