+++ packages/fieldSettings/settings/GroupToolSettings.ts
@@ -1,6 +1,5 @@
-import { JsonSchema } from '@jsonforms/core'
-import ToolsettingParts from '../mixins/ToolSettingParts'
-import { ToolSetting } from '@formswizard/types'
+import { ToolsettingParts } from '../mixins'
+import { ToolSetting, JsonSchema } from '@formswizard/types'
 
 const jsonSchema = {
   type: 'object',
